+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * SDL Text Input is activated by default
 * Fix input/hotkey mapping from config file (#302)
 * Check mask size in XISelectEvents
+* XI_RawMotion now uses raw inputs
 
 ## [1.3.5] - 2019-11-26
 ### Added